Refactoring: move utils out of main scripts
